clovers-agent/pyproject.toml
[project]
name = "clovers-agent"
version = "0.2.7"
description = ""
authors = [{ name = "KarisAya", email = "karisaya@foxmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "clovers>=0.5.1,<0.6.0",
    "clovers-client>=0.3.2,<0.4",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "torch>=2.5.0",
    "sentence-transformers>=5.3.0",
    "python-magic>=0.4.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clovers_agent"]
